requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
pymongo==4.5.0
bcrypt==4.0.1
python-jose==3.3.0
PyJWT==2.6.0
python-dotenv==1.0.0
pydantic==1.10.2